[SYCL] Fix more bugs with int-header and free function kernels
This time a more generic solution is proposed in form of a custom
visitor for template arguments.

The main trick is to _sometimes_ print template arguments in their
non-canonical form, which happens for:
- template arguments which were specialized
- template arguments which reference other template arguments

In all other cases we use canonical form which has all necessary
namespace qualifiers and accounts for type aliases.

This patch introduces some extra tests (specifically template template
one) which weren't a part of the original bug report which prompted this
patch. Those new tests are failing for now and added only for reference
and to highlight that there are still some bugs which will be addressed
in follow-up PRs.